Conflicts:
	ShareitClient/src/main/java/eg/iti/shareit/view/managedbeans/ItemDetailBean.java

+++ ShareitClient/src/main/java/eg/iti/shareit/view/managedbeans/ItemDetailBean.java
@@ -61,7 +61,7 @@ public class ItemDetailBean implements Serializable {
 
     @Inject
     private ListItemsBean listItemsBean;
-    
+
     @Inject
     private UserBean user;
 
@@ -97,7 +97,6 @@ public class ItemDetailBean implements Serializable {
         this.listItemsBean = listItemsBean;
     }
 
-    
     public ItemBean getItemBean() {
         return itemBean;
     }
@@ -106,8 +105,6 @@ public class ItemDetailBean implements Serializable {
         this.itemBean = itemBean;
     }
 
-
-    
     public List<String> getHashTags() {
         return hashTags;
     }
@@ -434,12 +431,20 @@ public class ItemDetailBean implements Serializable {
     }
 
     public String deleteItem() {
+
         try {
-            itemService.deleteSharedItem(item);
-            listItemsBean.getItems().remove(item);
-          //  itemBean.getItems().remove(item);
-            System.out.println("------------------- im delete service "+ itemBean.getItems().size());
-            return "items?faces-redirect=true";
+            if (itemService.isItemAvailable(id)) {
+                itemService.deleteSharedItem(item);
+                listItemsBean.getItems().remove(item);
+                //  itemBean.getItems().remove(item);
+                System.out.println("------------------- im delete service " + itemBean.getItems().size());
+                return "items?faces-redirect=true";
+            }
+            else {
+                FacesContext context = FacesContext.getCurrentInstance();
+                context.addMessage(null, new FacesMessage("Error", "you cannot delete the item"));
+                return "";
+           }
         } catch (ServiceException ex) {
             Logger.getLogger(ItemDetailBean.class.getName()).log(Level.SEVERE, null, ex);
             return "";
